refactor API models
Move all the model definitions that aren't exclusively used for internal
API operations (ie. all models related to user-facing endpoints) to
kernelci-core.

Move the UserGroup model to user_models.py
